Create MongoModel
There isn't much here in the way of functionality. At this point there is a `MongoModel` class, which is of type `MongoModelMetaClass`. The meta class is used primarily to associate the `Meta` class with subclasses; it checks for the required `collection` attribute and provides a default `field_map` attribute if none has been provided.

At this point in time the functionality of `MongoModel` is fairly limited. All that has been provided is the ability to get, set, and remove attributes that are meant to be part of the base document. There is no database interaction yet, so this commit is pretty useless, but it's a good point in time to capture in case the database interaction goes down a bad path.