Add basic connection behavior
This is by no means intended as the final state of connection to a database. It is, however, enough to get started building out the model functionality. A database connection can be opened through `connect()` and can be retrieved later on with `get_database()`.

I imagine that at some point I'll store multiple databases. I'd also like to add the ability to use existing connections. This would be useful for me as I use Flask-PyMongo in a lot of projects.