Add the ability to save
Just like all of previous other commits, this isn't necessarily meant to be the final version of this, but it's the first pass at saving data to the database.

In order to accomplish this, all models are given `default` as the name of the database--this is actually only done when no value is set in the model, but names other than default aren't supported yet--in the meta class, and the collection is associated with the model in its `__init__()`.

Right now `created` and `modified` are more or less forced upon each document. I'll probably add some settings to `Meta` to optionally override this behavior.