Remove ** from docstring
It appears that Sphinx sees `**` in a docstring as the beginning of a bit of boldface text. That makes sense because that is the delimiter used by reST for bold. While I had intended it to denote a keyword arguments dict.

Because Sphinx was throwing a warning for each unclosed `**`, I'm opting to remove them from the docstring.