Move association of collection to a decorator
When adding the saving ability in d8daa4e, I decided it would be a good idea to associate the collection object with the model at the time of instantiation. That way the collection would only be loaded up if it was actually being used.

While working on the document retrieval, however, I realized there was a major flaw in this approach: methods like `get()` would need to be class methods, and therefore not be used in conjunction with an instantiated object.

By creating the `requires_database` decorator, any method that will need to use the database can make sure the appropriate type, be it a class or an instance, will have a `db` attribute before trying to access it.