Restore db as an attribute without the use of a decorator
I really wasn't happy with the decorator approach for making sure that `db` was assigned to the `_meta` attribute of both classes and instances, especially because the decorator's arguments replaced any wrapped method's when generating Sphinx docs or using `help()`.

It is now back to being a property, this time associated with the model at creation time. This works because a lambda is used so that the collection is only accessed when needed. In order to achieve this, `property` is subclassed as `Property` with a custom `__get__()` method that will call the class method surrounding the lambda.

This is essentially a revert of 4263961.

Thanks to @vincent-kuwornu for helping me get this working.