Add ability to save only select fields
When updating documents, you don't usually want to replace the entire document when all you want to update are a subset of its fields. This is where the `$set` operator comes into play. The `save_fields()` method takes a list of fields to update with `$set`.

Unlike `save()`, the `modified` field will not be updated. I'm still on the fence about this one. My guess is that I'll change my mind before this ships.