Add ability to remove only select fields
When updating documents, you don't usually want to replace the entire document when all you want to remove are a subset of its fields. This is where the `$unset` operator comes into play. The `remove_fields()` method takes a list of fields to remove from the document with `$unset`.

Just as with ae0e7ea, the `modified` field will not be updated and I'm on the fence about whether or not that should be the case.