Change location of the meta document assignment
I think I learned an important lesson today: it appears that assigning an empty dictionary in `__new__()` is like using an empty dictionary as the default value for a parameter, the dictionary will persist throughout the life of the application.

`_meta.document` is now given its value in `__init__()` instead. This way each instance will have its own version of the dictionary.