Update save for upserted IDs
When upserting a new document, its `_id` is needed to continue working with the document through the instance that created it. How exactly this is done depends on whether or not the upsert was performed in safe mode.

When in safe mode, the result of the `lastError` command is returned. It contains the `updatedExisting` key which will be set to `False` for a new document. When that is the case, it also contains a key called `upserted` which is set to the `_id` of the new document.

When not in safe mode, the `_id` can be retrieved by querying for the same document through `find_one()`.

See https://groups.google.com/forum/?fromgroups=#!topic/mongodb-user/a0ZUwd1t3HE