Add tests for database interaction
These tests require the ability to connect to a MongoDB instance running on localhost. They will write to a database and collection called `test-simon`.

The database will be removed after each test is performed. This appears to be the slowest part of running the tests. I need to look into `unittest` a bit more to see if there's an equivalent to `tearDown` for after all tests have been run.