Improve database test speed
The slowest part of the database tests has been dropping the database after each test is complete. By utilizing `tearDownClass()` for this instead, the speed of these tests receives a much appreciated boost.

In order for `tearDownClass()` to be able to access the connection, the connection is now opened in `setUpClass()`.