Change location of internal document
f6f44aa wasn't a big enough change. Assigning a dictionary to `self._meta.document` at the time of instantiation is really no different than assigning it at creation time because `_meta` is associated with the class, not an instance. This means that any instances of a class with share the same dictionary.

    >>> m1 = SomeModel(a=1)
    >>> m2 = SomeModel(a=2)
    >>> print m1._meta.document
    {'a': 2}
    >>> print m2._meta.document
    {'a': 2}

This, obviously, is bad. :unamused:

I had originally placed `document` inside `_meta` because any attributes defined directly on the class are considered reserved names and can't be used for the underlying document. Hopefully most people don't use `_document` as a key name in their documents. I wanted to go one step further and use `__document` (to really emphasize that it's an internal attribute), but Python does same name mangling for attributes whose names begin with a dunder. I didn't want to complicate things by introducing this level of complexity.