Initial commit of find
This is by no means finished, but I've spend some time over the past couple of days writing and documenting the `MongoModel.find()` method and the `QuerySet` class. I'm happy with the direction I'm going in, so I want to make sure I don't lose what I have so far, especially in case I need to get back to it.

The final version of this functionality will have a much better commit message, but in the mean time, here's a :watermelon:.