Fix exception message for QuerySets with no model
When an `IndexError` is raised during a `QuerySet`'s `__getitem__()`, the name of the model class is used in the message to provide a more meaningful experience. Unfortunately a `QuerySet` can be created without a model. When this is the case, `QuerySet` will be used in place of the model name. Because this should only happen when a `QuerySet` is instantiated manually, this is more appropriate anyway.