Clean up limit() tests
The first change being made here is to use an index greater than the expected maximum index when filling the result caches. Failing to do this could cause a record not to load that would have otherwise, resulting in a incorrectly passing test.

The next change just cleans up a docstring to make it consistent with others.

The final change removes `test_limit_indexerror()`. `limit()` never throws an `IndexError`. This is more appropriately a test of `__getitem__()` and should be reimplemented as such.