Add tests for sort()
The tests check the following:

* sorting by a single key, ascending
* sorting by a single key, descending
* sorting by two keys, ascending
* sorting by two keys, descending
* sorting by two keys, first ascending, second descending
* sorting by two keys, first descending, second ascending

In order to properly test with multiple keys, each test needs to insert some extra documents because the original documents are contain a unique set of keys and values. Running these tests without these documents would result in all sorting being based on the first key of the sort.