Add a note to QuerySet iteration
Because I'm catching `StopIteration` in `_fill_to()`, I'm never going to reach 100% coverage in `query.py`. The exception will never be raised during tests because I already make sure the method doesn't try to include more documents that the `QuerySet` contains. This is, however, not necessarily the actual real world behavior since cursors are living things that can include changes to the collection at any time. If a document is removed from the database while iterating through the cursor to fill the cache, it is possible that the cursor's iterator would raise the exception.