Add new tests for _fill_to()
These tests are to ensure that the correct type is stored in the `QuerySet`'s result cache. When a model class is associated with the `QuerySet`, all items in the cache should be stored as instances of that class. When there is no class, all items should be stored as dictionaries.

I'm also adding a missing comment that should have been included for consistency.