Add a method to perform increment operations
This new `increment()` method can be used to perform atomic increments on the fields of an instance. It will perform the update and reload the changed fields.

Currently there is no protection in place for a field that contains an unsaved value. For example:

    obj.field = 17
    obj.increment('field')

This will result in `field` being `1` greater than the value in the database, not necessarily `18`.