Add a method to update fields
The `update()` method behaves very much like a shortcut for the `save_fields()` method. The difference is that `save_fields()` requires the values to be assigned previously, whereas `update()` ignores the values already associated with the fields and uses the values associated with the arguments passed in.

Because of the similarity, I may end up actually making `update()` a shortcut. I will probably run some tests to see which approach is more efficient.