Add a method for performing raw updates
The `raw_update()` method provides a way to perform an update using a MongoDB update document. The update will be passed directly to the database, so any operators used will be applied.

This method should be used carefully as it will perform the update exactly, potentially performing a full document replacement.

Also, for simple updates, it is preferred to use the `save()` or `update()` methods as they will usually result in less data being transferred back from the database.