Remove unneeded test helper methods
The `dummy_class_method()` and `dummy_method()` methods were added to `TestModel` when the `db` property was associated with classes and instances through the `requires_database` decorator. When the decorator was removed in d2cabe3, it was removed from these methods.

In actually, the methods should have been removed as well. They became obsolete.