Condense tests for db attribute into one class
There really is no point to checking for the `db` attribute in two separate tests. I'm moving them into one as it will cut out a little, although negligible, overhead.

The `isinstance` check is also being removed. This may seem odd now, but in a future revision the database will be mocked for this test so that a database connection won't be needed.

To make up for this, a test will be added into a class that does integration-like tests.