Fake the database collection
None of the tests in `TestBase` actually require a database connection. The only reason one would be needed is so `TestModel` can properly have its `db` attribute set. This check shouldn't rely on being able to connect to a database. By using a dummy dictionary, a fake collection can be added to the `connection` module's `__databases__` dictionary and used here.

A database name is associated with `TestModel` so as not to step on any other databases needed by other tests.