Rename remove to delete
This is essentially a revert of 30b18a1. Since the whole point of this library is to provide a layer of abstraction on top of PyMongo, it's method names don't need to be exactly the same as PyMongo's. And having one method that just aliases another clutters up the API.