Beef up upsert tests
I suspected that `find_one()` was returning the first document matching the upserted one when performing an unsafe upsert, not the newest one. To test this theory, I decided to make the upsert tests insert multiple documents, comparing _id values to make sure that each is unique.

As I suspected, the enhanced tests are failing.