Fix _id retrieval for upserted documents
When upserting a document with an unsafe write, retrieving that document's `_id` can be a bit tricky. Details can be found in e5490dd.

`find_one()` won't necessarily returned the inserted document. It will return *a* document matching the inserted one. The natural sort of the collection will determine which document is returned if there are multiple matching the query. By using a reverse sort on `_id`, `find_one()` will return the newest. This assumes that `_id` contains incrementing values, such as Object IDs.

This fixes the failing tests from 11d1b14.