Fix typo related to replica sets
:disappointed:

This typo resulted in an invalid argument being passed to `ReplicaSetConnection`. I was popping a key called 'replaceSet' from `settings` instead of 'replicaSet'. This resulted in a `ValueError` being raised as `replicaSet` is supposed to contain the name of the replica set, not a boolean like was being passed in.

This is why you write tests.