Begin process of caching connections
The original version of `connect()` only stored databases for later use. If two databases were used on the same server, two connections to the server were opened. The goal is to store both the databases and the connections so that they can be reused when necessary.

This is the first step in doing that. A `__connections__` dictionary and a `_get_connection()` method are being added to wrap around this functionality. Currently the logic has been moved from `connect()` to `_get_connection()`. The next step is to begin refactoring the logic as many of the parameters being passed to `_get_connection()` are specific to the database, not the connection.

The tests are being updated here to mock `_get_connection()` instead of PyMongo's `Connection`. Both will probably happen in a future iteration.