Rework connect() and _get_connection()
This is the next step after 5572b44.

`_get_connection()` now does a lot of the work that `connect()` used to it. It now parses URIs. In addition to now only taking the parameters that it needs, it gives back anything it doesn't need from a parsed URI.

The tests are also being refactored to test both the `connect()` method and the `_get_connection()` method. The test for `connect()` mocks `_get_connection()` and the tests for `_get_connection()` mock `Connection` and `ReplicaSetConnection`.

Also, a change is being made to the docstring of `connect()` to make it consistent with the one being added for `_get_connection()`.