Add method to parse keyword argument keys
One of the most useful features of MongoDB is its ability to handle nested documents. At this point in time, Simon can only handle them by assigning a dictionary to a keyword argument or attribute.

The goal of the `parse_kwargs()` method is to change it. Once it's implemented in other methods, the documents can be retrieved with calls such as `SomeModel.find(a__b=1)` and new documents can be created along the lines of `SomeModel(a__b=1)`.

I debated placing this method in a utils module, but I thought that would be overkill for one method. If something else comes along, I'll move it there. I'll also end up renaming the test_miscellaneous module to test_utils at that time.