Fix TypeError
If someone decided to use `get_database()` before connecting to one, a `TypeError` would be raised because `__databases__` is originally set to `None`, and `None` is not iterable.

This is an error that would have been presenting itself in the `ConnectionError` test for a while now. I've had this line uncommitted for a while as I debated about making the check for `None` separate. In the end I decided that both problems are remedied in the same one, so have one exception point is appropriate.