Add a method to handle mapping fields
There are a lot of occasions that call for mapping field names. Most of `MongoModel`'s methods implement some code to handle mapping attributes to keys. Unfortunately this doesn't help with the keys of embedded documents. Some methods failed to implement even this.

The aim of `map_fields()` is to provide a way to to handle mapping not just the top level keys, but also the nested ones as well. It will also handle calls to `parse_kwargs()`, allowing for a one-line implementation by methods that need to map fields.