Add missing tests
My coverage report was showing 100% coverage for all of the modules. This seemed odd for two reasons:

1. `query.py` contains a catch for a `StopIteration` that is included as a "just in case" and is not covered by any test.
2. I had just noticed there were no tests for `MongoModel.find()`.

I removed `.coverage` and lo and behold there were quite a few tests missing.

These tests should cover all of those missing things. All remaining uncovered lines are item 1 above, exceptions that knowingly pass silently, and a couple of lines of code that appears to be unnecessary, but I'm not yet brave enough to remove.