Add protection for keys beginning and/or ending with '__'
Keys beginning or ending with `__` are valid and should not be replaced by keys beginning or ending with `.`. While the assertions in `test_map_fields()` hold true without any changes to `map_fields()`, this is not the case for the assertions in `test_map_fields_flattened_keys()`.

A change has been made to `map_fields()` to ensure that `__`s at the beginning and/or end of a key will not be replaced.