Add support for mapped and nested fields to remove_fields()
Continuing with the trend of updating all of the methods that write to the database to support mapped fields, `remove_fields()` is now calling `map_fields()`. In order to properly accommodate mapped fields, the method also utilizes `remove_nested_key()` to remove keys nested in the internal document.

If support is ever added to treat embedded documents as anything more sophisticated that a dictionary, this method will need to be updated.