Truncate Python datetimes to match BSON
Python calculates datetimes using microseconds. BSON calculates them using milliseconds. When updating `test_save_insert()` to compare not just the dictionary keys but the values too, the test began to fall. The difference in the two dictionaries was the precision of the datetime values.

By removing Python's extra precision, not only does the modified test pass, but the data becomes more consistent. Rather than having one modified date associated with the object and another stored in the database, both values will always be the same.