Replace xrange() with range()
I hope to some day get Python 3 support working. I've already been using `.items()` instead of `.iteritems()` to iterate through dictionaries. Using `range()` is the next step. Most of the occurrences were in tests, so only the reference in the `QuerySet.__iter__()` method needs to be changed in the package.

Of course Python 3 uses a completely different syntax for metaclasses. That change is for another time.