Fix connecting to a replica set
tl;dr: `ReplicaSetConnection` wants a field called `hosts_or_uri`, not `host_or_uri` and the `replicaSet` parameter needs to be a string, not a boolean.

I ran some (not yet committed) unit tests that connect to real databases, including a replica set. All hell broke loose.

First I received `ConfigurationError` because `host_or_uri` is an unknown option. The parameter is actually called `hosts_or_uri`.

Then I received a `TypeError` because I was passing a boolean for the `replicaSet` parameter. The value needs to contain the name of the replica set. It's not an indicator to the `ReplicaSetConnection`--the name really should have been a hint--that it's supposed to connect to a replica set.

These changes address these two issues and update the tests associated with them. Additionally, the `__connections__` dictionary uses the name of the replica set instead of the port number. There's a note explaining why this could get problematic.