Add better support for field mapping to __getattr__()
The `__getattr__()` method now supports nested keys. This has two important uses. Firstly, `getattr()` can be used to request a nested key from the internal document using the `.` syntax. Secondly, attributes of the instance can be used to access a nested key from the internal document using the `__` syntax.

As part of this change, the test for `__getattr__()` is being refactored a bit. The method is probably more accurately tested by assigning keys to the internal document directly rather than through `__setattr__()`.