Add field mapping to save_fields()
It seems like the task of introducing field mapping is going to take forever. But the list of outstanding methods is definitely getting shorter.

Here the `save_fields()` method not only gets support for mapped fields, but it also gets support for nested fields. It was really the impetus for 1d0d519. I feel like a lot of good stuff has come out of this process.