Allow automatic timestamps to be disabled
The goal of Simon is to have as few restrictions on the data as possible. That goes both ways. As originally planned out, `MongoModel.save()` would add `created` and `modified` dates to documents. This may not be something that everyone wants.

Adding the `auto_timestamp` setting to the `Meta` class also this behavior to be disabled at the model level. While the default behavior remains unchanged, set `auto_timestamp` to `False` will prevent the timestamps from automatically being added to documents.