Allow models to perform all writes in safe mode
The `safe` attribute is being introduced to allow all writes (and deletions) by a model to be performed in safe mode. This will override the `safe` parameter for  any method defined as part of`MongoModel`.

I want to add an additional test to make sure each method respects the behavior correctly, but I am short on time and want to get this done before I lose the thought.