Add comparison operators
MongoDB supports several comparison operators when performing queries. This is my first pass at adding support for them. This adds:

* $gt, $gte
* $lte, $lte
* $ne
* $in, $nin
* $exists

These can be used in queries by appending `__[operator_without_$]` to the end of a key. Such keys can be passed to the `map_fields()` method with `with_comparisons` set to `True` and it will adjust the query syntax.

It's probably a good idea to add some additional tests to make sure the `get()` and `find()` methods support these correctly.