Add support for logical operators
MongoDB supports the logical operators `$and` and `$or`. I wanted to expose them without requiring developers to write the raw query spec. I turned to Django's ORM for the influence in how to do this: the `Q` object.

`simon.query.Q` objects can be combined through `&` and `|` to build queries using the `$and` and `$or` operators. A single `Q` object by itself will not use either operator. Chaining multiple objects together with the same logical Python operator with add them to the same list of filters for the logical MongoDB operator. Chaining them together with different Python operators will result in nested MongoDB operators.

`Q` objects can be passed to `MongoModel.find()` and `MongoModel.get()` to filter queries. `find()` and `get()` will extend the filters passed through the kwargs. From there, `map_fields()` will handle mapping any keys contained in the model's `field_map`. I had wanted this step to happen when creating each `Q` object. Unfortunately, though, the model class is needed in order to perform the mapping. I didn't want to have to include the class in each `Q` constructor; I left the syntax used by Django was more natural.

At this time `$nor` and `$not` are not yet supported.