Rename MongoModel to Model
When I first started working on this project, I opted to name the primary class `MongoModel` to avoid naming conflicts with other libraries that use `Model`. I've since changed my mind and am renaming it to `Model` to be consistent with how other libraries work.

As an added bonus, less typing is needing to define subclasses.