Add support for the $not operator
MongoDB allows for `$not` to be inserted into queries. Simon models can now be queried using the `$not` operator by adding it to the keyword argument name between the key and the operator (e.g., a__not__gt=1).

I'm also changing the operators in the `map_fields()` docstring to include the `$`. I've elected to do this because that is how they are referred to in the MongoDB documentation.