Add helper method to validate points
All of the geospatial helper methods--with the exception of `within()`--make sure the points provided for their arguments are valid. This caused the same checks to appear all over the place. By adding a helper method, a lot of the code can be removed.

As a result of this change, `query.near()` now raises `ValueError`. This behavior should have existed from the beginning, but I didn't add `ValueError` to the geospatial methods until after `near()` was complete.