Update docstrings
I set out to add `*` and `**` to all of the args and kwargs arguments. In addition, `(optional)` should not be included with the description for all optional arguments.

While doing this I noticed that the docstring for `Model.update()` used the description of `fields` from `Model.save_fields()` for its own `fields` argument.