Remove set() from test of distinct()
Taking the result from `distinct()` and converting it to a set will guarantee that only unique values are in the set, but there is no way to know if any duplicates were removed by `distinct()` or by `set()`. This makes for a poor test.

I believe the reason I originally used `set()` was because `distinct()` doesn't guarantee the order in which the values will appear. I needed a way to reliably test for the values. Testing the length of the result and making sure each expected value is in the result seemed like more work than necessary.

Using `sorted()` to control the order of the result allows the values returned by `distinct()` to be compared to the expected values--just as `set()` did--but without altering the list.