Add membership test method to Model
A `__contains__()` method is being added to `Simon.Model` to make it possible to test for a key in a document through the membership test operators (`in` and `not in`). While it is still possible to use `hasattr()`, there are times when it may more aesthetic to use `in` (e.g., when checking for both the presence of an embedded document and one of its keys).