Move geospatial helpers to their own module
The original intention of the `query` module was to house the `QuerySet` class. Over time it grew to contain `Q` and the geospatial helper methods as well. While I'm opting to leaving `Q` in the `query` module I decided it was more appropriate for the geospatial methods to be placed in their own module.

By adding a `geo` module, all methods associated with geospatial queries can be found and viewed without being mixed in with the unrelated `Q` and `QuerySet` classes.