Add method for converting to Object IDs
When querying by a document's (default) `_id`, it is important that the value is handled as an Object ID and not just the string representation of the ID. Originally this was checked in `find()` and `get()` directly. The checks gets more complex when you need to handle comparisons and not just direct lookups.

The `guarantee_object_id()` method is being added to handle all necessary checks and conversions. Not only does it reduce the number of places the logic will need to exist, it also means the functionality can be unit tested outside of the context of querying the database.