Add support for mapped fields to membership test
In order to properly perform a membership test on a `Model`, key names need to be mapped. For example, without this change, the statement `'id' in m` would evaluate to `False`.

In addition to a new test to test this behavior for `__contains__()`, tests are being added for `__delattr__()`, `__getattr__()`, and `__setattr__()` as well.