Begin support for PyMongo 2.4
With the release of PyMongo there are two significant changes. First, `Connection` and `ReplicaSetConnection` have been marked as deprecated in favor of the newly added `MongoClient` and `MongoReplicaSetClient`.

The second change is how the new classes handle write concern. By default write concern is now enabled and can be controlled through the `w` parameter. At this time I'm still using `safe`, but since this, too, is now deprecated I'll probably make the change to `w` in the near future.