Continue the trend of adding __repr__(), __str__(), and __unicode__() methods
While it's a good idea to define methods like `__str__()` and `__unicode__()` for each model, it's a good idea for `Model` to provide a base set. `__repr__()` can probably be left as is.

In order to properly test the methods, I'm adding a second test model for `TestBase` to use.

While adding these methods to `Model` I figured `Meta` should have `__unicode__()` as well. In addition to adding a test for `__unicode__()`, I'm also updating the docstrings for `Meta`'s `__repr__()` and `__str__()` tests to be consistent with the others.