WIP: Mock database calls
Because PyMongo handles the actual database interaction, I just need to ensure that I'm calling the PyMongo methods with the correct parameters. I plan to mock as many of these methods as possible.

Normally I wouldn't commit in the middle of a task, but I'm going to be away from my computer for a bit and I want to make sure this work doesn't get lost.

References #3