WIP: Mock database calls
Continuation of fcc3ca7

Because PyMongo handles the actual database interaction, I just need to ensure that I'm calling the PyMongo methods with the correct parameters. I plan to mock as many of these methods as possible.

Normally I wouldn't commit in the middle of a task, but I'm going to be away from my computer for a bit (yay sleep) and I want to make sure this work doesn't get lost.

References #3