WIP: Mock database calls
Continuation of fcc3ca7

Because PyMongo handles the actual database interaction, I just need to ensure that I'm calling the PyMongo methods with the correct parameters. I plan to mock as many of these methods as possible.

Normally I wouldn't commit in the middle of a task, but since that pattern has already been established for this task I've decided to continue taking periodic snapshots. `QuerySet._fill_to()` is the method that I've been having the most trouble figuring out how to integrate Mock into. I'm using this as a chance to make sure everything else is in place before attempting to tackle it. Some class-level stuff still needs to be addressed, but I'll take care of that everything `_fill_to()` is done.

References #3