Test housekeeping
Tests now do a better job of cleaning up after themselves.

While making these changes, `TestMetaClass.test_db()` began failing. Turns out it was taking advantage of `TestConnection` cleaning up *before* its tests instead of after. The type check on the `db` property is being removed. This allows the test to pass.

It's important to note that I'm not removing this check simply to get the test to pass again. `test_db()` already ensures that references the `db` property will make a call to `connection.get_database()`. As far as unit testing goes, this should be enough for `db`. It's up to any unit tests for `get_database()` to make sure that it's doing the right thing.

This is also why `TestDatabase.test_db_attribute()` is being removed. Because the `db` property is already being tested in `TestMetaClass` there's no need to test it again in `TestDatabase`.