Add integration tests
These new integration tests are designed to make sure that the code in the `connection` module can properly connect to a MongoDB database. They can be invoked by executing `make integration` at the command line.

The tests are incorporated in the tox suite. I noticed, however, that the tests were failing to finish when run using PyMongo 2.3. Because of this the integration tests won't run with that version of the driver. I hope to find the cause of this and address it.

The key to the integration tests is a local-only module called `integration_settings`. It should contain the values needed to connect to the remote databases used for testing. The integration tests that rely on these settings are designed to be skipped if the necessary setting isn't found.

The new .gitignore file is designed to prevent these settings from accidentally being added to the repository.