Add required_fields option to Meta
The first step is adding support for required fields is to add an option to the `Meta` class. The option defaults to `None` but can be set to either a single field or a list (or tuple) of fields. If a single field is provided, it will be converted to a tuple so that all `_meta.required_fields` are treated the same.

References #2