Switch connect() kwarg to replica_set
When I first wrote `connection.connect()` I opted to use `replicaSet` to match the syntax used by the URI connection strings. After using it a few times I've decided it doesn't feel Pythonic to me. The kwarg is changing to `replica_set`. To maintain backward compatibility, `replicaSet` will be checked as well, but only in the absence of `replica_set`.

I would have loved to have written this as

    replica_set = kwargs.pop('replica_set', None)
    kwargs.pop('replicaSet', None)
    if replica_set is None:
        replica_set = _

but alas that only works in the shell.